+++ app/javascript/engine/control/deploy.ts
@@ -153,7 +153,6 @@ function collapseRows(hexes: DeployHexes): DeployHexes {
     len += 1
     if (rc0[i][0] === nn[0] && typeof nn[1] === "number" && typeof rc0[i-1][1] === "number" &&
         rc0[i][1] === (rc0[i-1][1] as number) + 1) { continue }
-    console.log("bye")
     rc1.push([nn[0], len < 2 ? nn[1] : `${start}-${rc0[i - 1][1]}`])
     start = rc0[i][1]
     nn = rc0[i]