+++ spec/models/utility/scenario_spec.rb
@@ -183,7 +183,7 @@ RSpec.describe Utility::Scenario do
       constants = Scenarios.constants.reject do |k|
         %i[Base Scenario999 Scenario0TT].include?(k)
       end
-      expect(constants.length).to be == 130
+      expect(constants.length).to be == 131
 
       # If any of these change, scenario MUST to be updated with a new version,
       # then update test with new version/checksum
@@ -225,7 +225,7 @@ RSpec.describe Utility::Scenario do
       expect(Utility::Scenario.checksum("034")).to be == "0.1p-365a0855cec368d34c9475f5631de49c"
       expect(Utility::Scenario.checksum("035")).to be == "0.1p-b9cb4c16a792186fbfd6147f69efa836"
       expect(Utility::Scenario.checksum("036")).to be == "0.1p-1ae949168796a5f726d3faef8c116c58"
-
+      expect(Utility::Scenario.checksum("037")).to be == "0.1p-b80f5f8c2725d44c6310364e7b7ad451"
       expect(Utility::Scenario.checksum("038")).to be == "0.1p-394d28ab9b64381988a430cea6211944"
       expect(Utility::Scenario.checksum("039")).to be == "0.1p-d58f26cf8076394e8b04cc69058d8ddb"
       expect(Utility::Scenario.checksum("040")).to be == "0.1p-fbd794ebb31d01d242fd1e938c6171d9"